# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Minimalist Identity \u0026 Immutable Activity Engine",
        "title": "MoolRoop Identity Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/activities": {
            "post": {
                "description": "Creates an append-only, immutable activity log event for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Log an immutable activity record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-supplied correlation request ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "description": "Activity payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.LogActivityPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_gen_v1.ActivityRecord"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Unique correlation request ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activities/{user_id}": {
            "get": {
                "description": "Retrieves all historical activity records for the specified user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "List immutable activity history for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-supplied correlation request ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_moolroop_gen_v1.ActivityRecord"
                            }
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Unique correlation request ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles": {
            "post": {
                "description": "Creates a new user profile with minimal PII (full name and email only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Create a new user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-supplied correlation request ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "description": "Create Profile Request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.CreateProfilePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_gen_v1.UserProfile"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Unique correlation request ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profiles/{id}": {
            "get": {
                "description": "Retrieves profile information by user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Get user profile by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-supplied correlation request ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_gen_v1.UserProfile"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Unique correlation request ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Incrementally update profile fields without rewriting full entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Patch an existing user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Optional client-supplied correlation request ID",
                        "name": "X-Request-ID",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to patch",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.PatchProfilePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_gen_v1.UserProfile"
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "Unique correlation request ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moolroop_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_moolroop_gen_v1.ActivityRecord": {
            "type": "object",
            "properties": {
                "action_type": {
                    "type": "string"
                },
                "activity_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "logged_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_moolroop_gen_v1.UserProfile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_moolroop_internal_model.CreateProfilePayload": {
            "type": "object",
            "required": [
                "email",
                "full_name"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "alice@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Alice Smith"
                }
            }
        },
        "github_com_moolroop_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "user profile not found"
                }
            }
        },
        "github_com_moolroop_internal_model.LogActivityPayload": {
            "type": "object",
            "required": [
                "action_type",
                "user_id"
            ],
            "properties": {
                "action_type": {
                    "type": "string",
                    "example": "PROFILE_UPDATE"
                },
                "description": {
                    "type": "string",
                    "example": "User updated profile contact info"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_moolroop_internal_model.PatchProfilePayload": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "alice.new@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Alice Cooper"
                }
            }
        },
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. This field is\nthe nanosecond portion of the duration, not an alternative to seconds.\nNegative second values with fractions must still have non-negative nanos\nvalues that count forward in time. Must be between 0 and 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must\nbe between -62135596800 and 253402300799 inclusive (which corresponds to\n0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).",
                    "type": "integer"
                }
            }
        }
    }
}